Reuse code with AbstractWithMetaObjectTypeFieldResolver
